Introduce rename tasks and a task manager
During recursive renames, problems occur if a directory is renamed, and
then files or directories inside that directory are trying to be renamed,
because they don't exist anymore at that path.

This commit splits the rename process in tasks, taking care that the
right rename order is preserved. Also, the success objects need to be
updated, so that the new GFile reference correctly reflects later
prefix path renames.