Handle byte-compilation during install manually
Automake's Python support creates byte-compiled versions of Python
modules during install time, with and without optimization.

However, as of Python 3.2, the handling of byte-compiled files has
changed, see PEP 3147. The files that Automake creates are now ignored,
and in the normal case where the user doesn't have write permissions to
the installation directory, this results in byte-compilations of every
application startup (and thus a significant startup performance penalty).

Thus, until an updated Automake is used in this project, the
byte-compilation procedure during the install stage is handled manually.